Moved IEchoGrain
