+++ src/Outkeep.Grains.Interfaces/OutkeepGrainFactoryExtensions.cs
@@ -1,5 +1,5 @@
-﻿using Outkeep.Grains;
-using Outkeep.Grains.Caching;
+﻿using Outkeep.Grains.Caching;
+using Outkeep.Grains.Echo;
 using System;
 
 namespace Orleans
